Fixed sizing on larger displays
I used @media to make everything look the right size when it gets scaled up onto a larger display. I also have a question that probably has a really simple answer: Would it be possible to make "Raffle Progress Bar!" bold? I feel like making it a little smaller but then bolding it would make it work better with the rest of the page.