Removed link
It seems as though the link isn't working because it's a list. Everywhere else, it works, but I'm not sure how to make it work in the list. We don't need it  linked anyway; the page should already be opened by the people who set up the Hour of Code.