Updated styles to maximize image size
The headphone image is in a vertical orientation so should be given the class vertical. Other than that, you want all of the images other than the first place prize to have a similar visual impact. So I did my best to make them all a similar size. Feel free to tweak the widths as you see fit.